SFR-711 more improvements
