--- src/react-components/src/components/02-molecules/search-bar/ParamSearchBar.stories.tsx
@@ -1,13 +1,0 @@
-import * as React from "react";
-
-import ParamSearchBar from "./ParamSearchBar";
-import { action } from "@storybook/addon-actions";
-import { withKnobs, text } from "@storybook/addon-knobs";
-
-
-export default {
-  title: "ParamSearchBar",
-  component: ParamSearchBar,
-};
-
-export const basicSearch = () => <ParamSearchBar searchBarId="searchBarId" dropdownId="dropdownId" dropdownOptions={["hi there", "bye there"]} searchHandler={action("searched")} dropdownBlurHandler={action("selectChanged")}/>;
