--- src/react-components/src/components/02-molecules/search-bar/SearchBar.stories.tsx
@@ -1,12 +1,0 @@
-import * as React from "react";
-
-import SearchBar from "./SearchBar";
-import { action } from "@storybook/addon-actions";
-
-export default {
-  title: "SearchBar",
-  component: SearchBar,
-};
-
-export const basicSearch = () => <SearchBar id="searchbar" changeHandler={action("changed")} searchHandler={action("searched")}/>;
-export const searchWithPlaceholder = () => <SearchBar id="searchbar" placeholderText={"Placeholder"} searchHandler={action("searched")}/>;
