+++ src/react-components/src/index.ts
@@ -1,6 +1,6 @@
 export { default as Button } from "./components/01-atoms/Button/Button";
 export { default as BodyText } from "./components/01-atoms/Text/Text/BodyText";
-export { default as Breadcrumb } from "./components/02-molecules/menus/Breadcrumbs/Breadcrumbs";
+export { default as Breadcrumb } from "./components/02-molecules/Menus/Breadcrumbs/Breadcrumbs";
 export { default as UnderlineLink } from "./components/01-atoms/Links/UnderlineLink";
 export { default as IconLink } from "./components/01-atoms/Links/IconLink";
-export { default as ParamSearchBar } from "./components/02-molecules/search-bar/ParamSearchBar";
+export { default as ParamSearchBar } from "./components/02-molecules/SearchBar/ParamSearchBar";
